move tensor product in dedicated files
